Apply Anton's suggestions
See comments:
* https://github.com/jesyspa/kotlin/pull/172#discussion_r1377630198
* https://github.com/jesyspa/kotlin/pull/172#discussion_r1377629704
* https://github.com/jesyspa/kotlin/pull/172#discussion_r1377628258
