+++ plugins/formal-verification/formver.plugin/src/org/jetbrains/kotlin/formver/ViperPoweredDeclarationChecker.kt
@@ -54,8 +54,8 @@ class ViperPoweredDeclarationChecker(private val session: FirSession, private va
             val verifier = Verifier()
             val onFailure = { err: VerifierError ->
                 val source = err.position.unwrapOr { declaration.source }
-                with(VerifierErrorInterpreter) {
-                    reporter.reportOn(source, err, config.errorStyle, context)
+                with(VerifierErrorInterpreter()) {
+                    reporter.reportVerifierError(source, err, config.errorStyle, context)
                 }
             }
