+++ plugins/formal-verification/testData/diagnostics/bad_contracts/returns_booleans.fir.diag.txt
@@ -9,3 +9,15 @@ method global$fun_incorrectly_returns_false$fun_take$$return$T_Boolean()
 }
 
 /returns_booleans.kt:(183,196): warning: Function may return a false value.
+
+/returns_booleans.kt:(264,288): info: Generated Viper text for incorrectly_returns_true:
+method global$fun_incorrectly_returns_true$fun_take$$return$T_Boolean()
+  returns (ret: Bool)
+  ensures ret == false
+{
+  ret := true
+  goto label$ret$0
+  label label$ret$0
+}
+
+/returns_booleans.kt:(325,339): warning: Function may return a true value.
