+++ plugins/formal-verification/testData/diagnostics/bad_contracts/returns_booleans.kt
@@ -9,3 +9,11 @@ fun <!VIPER_TEXT!>incorrectly_returns_false<!>(): Boolean {
     return false
 }
 
+@OptIn(ExperimentalContracts::class)
+fun <!VIPER_TEXT!>incorrectly_returns_true<!>(): Boolean {
+    contract {
+        <!UNEXPECTED_RETURNED_VALUE!>returns(false)<!>
+    }
+    return true
+}
+
