+++ plugins/formal-verification/testData/diagnostics/bad_contracts/returns_not_null.fir.diag.txt
@@ -1,0 +1,12 @@
+/returns_not_null.kt:(121,133): info: Generated Viper text for returns_null:
+method global$fun_returns_null$fun_take$$return$NT_Int()
+  returns (ret: dom$Nullable[Int])
+  ensures dom$Type$isSubtype((dom$TypeOf$typeOf(ret): dom$Type), dom$Type$special$Nullable(dom$Type$Int()))
+  ensures ret != (dom$Nullable$null(): dom$Nullable[Int])
+{
+  ret := (dom$Casting$cast((dom$Nullable$null(): dom$Nullable[dom$Unit]), dom$Type$special$Nullable(dom$Type$Int())): dom$Nullable[Int])
+  goto label$ret$0
+  label label$ret$0
+}
+
+/returns_not_null.kt:(167,183): warning: Function may return a null value.
