--- plugins/formal-verification/testData/diagnostics/bad_contracts/user_friendly.fir.diag.txt
@@ -1,48 +1,0 @@
-/user_friendly.kt:(185,210): info: Generated Viper text for non_nullable_returns_null:
-method global$fun_non_nullable_returns_null$fun_take$T_Int$return$T_Int(local$x: Int)
-  returns (ret: Int)
-  ensures false
-{
-  ret := local$x
-  goto label$ret$0
-  label label$ret$0
-}
-
-/user_friendly.kt:(249,262): warning: Function may return a non-null value.
-
-/user_friendly.kt:(326,338): info: Generated Viper text for returns_null:
-method global$fun_returns_null$fun_take$$return$NT_Int()
-  returns (ret: dom$Nullable[Int])
-  ensures dom$Type$isSubtype((dom$TypeOf$typeOf(ret): dom$Type), dom$Type$special$Nullable(dom$Type$Int()))
-  ensures ret != (dom$Nullable$null(): dom$Nullable[Int])
-{
-  ret := (dom$Casting$cast((dom$Nullable$null(): dom$Nullable[dom$Unit]), dom$Type$special$Nullable(dom$Type$Int())): dom$Nullable[Int])
-  goto label$ret$0
-  label label$ret$0
-}
-
-/user_friendly.kt:(372,388): warning: Function may return a null value.
-
-/user_friendly.kt:(455,480): info: Generated Viper text for incorrectly_returns_false:
-method global$fun_incorrectly_returns_false$fun_take$$return$T_Boolean()
-  returns (ret: Bool)
-  ensures ret == true
-{
-  ret := false
-  goto label$ret$0
-  label label$ret$0
-}
-
-/user_friendly.kt:(517,530): warning: Function may return a false value.
-
-/user_friendly.kt:(598,622): info: Generated Viper text for incorrectly_returns_true:
-method global$fun_incorrectly_returns_true$fun_take$$return$T_Boolean()
-  returns (ret: Bool)
-  ensures ret == false
-{
-  ret := 1 == 1
-  goto label$ret$0
-  label label$ret$0
-}
-
-/user_friendly.kt:(659,673): warning: Function may return a true value.
