+++ plugins/formal-verification/tests-gen/org/jetbrains/kotlin/formver/plugin/runners/FirLightTreeFormVerPluginDiagnosticsTestGenerated.java
@@ -88,15 +88,15 @@ public class FirLightTreeFormVerPluginDiagnosticsTestGenerated extends AbstractF
         }
 
         @Test
-        @TestMetadata("returns_null.kt")
-        public void testReturns_null() throws Exception {
-            runTest("plugins/formal-verification/testData/diagnostics/bad_contracts/returns_null.kt");
+        @TestMetadata("returns_not_null.kt")
+        public void testReturns_not_null() throws Exception {
+            runTest("plugins/formal-verification/testData/diagnostics/bad_contracts/returns_not_null.kt");
         }
 
         @Test
-        @TestMetadata("user_friendly.kt")
-        public void testUser_friendly() throws Exception {
-            runTest("plugins/formal-verification/testData/diagnostics/bad_contracts/user_friendly.kt");
+        @TestMetadata("returns_null.kt")
+        public void testReturns_null() throws Exception {
+            runTest("plugins/formal-verification/testData/diagnostics/bad_contracts/returns_null.kt");
         }
     }
